refactor carousel code, add and update readme file
