--- README.old.md
@@ -1,2 +1,0 @@
-# ruby-resorts-frontend
-Frontend app for Ruby Resorts web page.
