+++ src/App.js
@@ -2,7 +2,7 @@ import './App.css';
 import { Routes, Route } from 'react-router-dom';
 import { Box } from '@mui/material';
 import NavigationPanel from './components/navigation/navigationPanel';
-import MainPage from './components/mainPage';
+import MainPage from './components/mainpage/mainPage';
 import ReservePage from './components/reservePage';
 
 const App = () => (
